+++ pyproject.toml
@@ -9,7 +9,8 @@ dependencies = [
     "numpy",
     "pandas",
     "plotly",
-    "streamlit"
+    "streamlit",
+    "click"
 
 ]
 requires-python = ">=3.11"
@@ -26,3 +27,6 @@ classifiers = [
 
 [project.urls]
 Homepage = "https://github.com/adhimmulia/ExImReservoirSimulation"
+
+[project.scripts]
+eximresim = "eximreservoirsimulation.main:entry_point"
